% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcValidNitrogenAtmosphericDeposition.R
\name{calcValidNitrogenAtmosphericDeposition}
\alias{calcValidNitrogenAtmosphericDeposition}
\title{calcValidNitrogenAtmosphericDeposition}
\usage{
calcValidNitrogenAtmosphericDeposition(datasource = "CEDS")
}
\arguments{
\item{datasource}{Bodirsky for own calculations based on Dentener et al, CEDS et al and self-calcualted emissions}
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description.
}
\description{
Validation Script for Atmospheric nitrogen deposition
}
\examples{

\dontrun{ 
calcOutput("ValidNitrogenAtmosphericDeposition")
}

}
\author{
Benjamin Leon Bodirsky
}
