% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcValidNitrogenBudgetNonagland.R
\name{calcValidNitrogenBudgetNonagland}
\alias{calcValidNitrogenBudgetNonagland}
\title{calcValidNitrogenBudgetNonagland}
\usage{
calcValidNitrogenBudgetNonagland(datasource = "Bodirsky")
}
\arguments{
\item{datasource}{Bodirsky for own calculations}
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description.
}
\description{
Validation Script for Nitrogen Budgets on Non-Agricultural Land
}
\examples{

\dontrun{ 
calcOutput("ValidNitrogenBudgetNonagland")
}

}
\author{
Benjamin Leon Bodirsky
}
