% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcValidNitrogenBudgetOcean.R
\name{calcValidNitrogenBudgetOcean}
\alias{calcValidNitrogenBudgetOcean}
\title{calcValidNitrogenBudgetOcean}
\usage{
calcValidNitrogenBudgetOcean(datasource = "Bodirsky")
}
\arguments{
\item{datasource}{Bodirsky for own calculations}
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description.
}
\description{
Validation Script for Nitrogen Budgets for Oceans
}
\examples{

\dontrun{ 
calcOutput("ValidNitrogenBudgetOcean")
}

}
\author{
Benjamin Leon Bodirsky
}
