% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcValidNutrientWasteBudget.R
\name{calcValidNutrientWasteBudget}
\alias{calcValidNutrientWasteBudget}
\title{calcValidNutrientWasteBudget}
\usage{
calcValidNutrientWasteBudget(datasource = "Bodirsky", nutrient = "nr")
}
\arguments{
\item{datasource}{Bodirsky for own calculations}

\item{nutrient}{The nutrient in which the results shall be reported.}
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description.
}
\description{
Report waste flows of nutrients from differents sources, including Household waste, slaughterwaste and Processingwaste
}
\examples{

\dontrun{ 
calcOutput("ValidNutrientWasteBudget")
}

}
\seealso{
\code{\link{calcValidNutrientBudgetSewage}}
}
\author{
Benjamin Leon Bodirsky
}
