% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcValidPlanetaryBoundariesNitrogen.R
\name{calcValidPlanetaryBoundariesNitrogen}
\alias{calcValidPlanetaryBoundariesNitrogen}
\title{calcValidPlanetaryBoundariesNitrogen}
\usage{
calcValidPlanetaryBoundariesNitrogen(datasource = "Rockstroem2009")
}
\arguments{
\item{datasource}{datasource to compare to. Historical trajectories of the indicators, or the planetary boundary by the Rockstroem Paper 2009 or the Steffen et al paper 2015.}
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description.
}
\description{
Provides comparison for the planetary boundary indicators for nitrogen
}
\examples{

\dontrun{ 
calcOutput("ValidPlanetaryBoundariesNitrogen",datasource="Rockstroem2009",aggregate=FALSE)
}

}
\author{
Benjamin Leon Bodirsky
}
