% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcValidSNUpE.R
\name{calcValidSNUpE}
\alias{calcValidSNUpE}
\title{calcValidSNUpE}
\usage{
calcValidSNUpE(datasource = "Bodirsky")
}
\arguments{
\item{datasource}{Bodirsky for own calculations, Lassaletta2014 for a country dataset from 
Lassaletta, L., G. Billen, B. Grizzetti, J. Angalde, and J. Garnier. 2014. 
50 Year Trends in Nitrogen Use Efficiency of World Cropping Systems: The Relationship between Yield and Nitrogen Input to Cropland.
Environmental Research Letters.}
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description.
}
\description{
Validation Script for Soil Nitrogen Uptake Efficiency (see Bodirsky 2012 for a definition)
}
\author{
Benjamin Leon Bodirsky
}
