% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullValidHistoricalNitrogenBudgets.R
\name{fullVALIDHISTORICALNITROGENBUDGETS}
\alias{fullVALIDHISTORICALNITROGENBUDGETS}
\title{fullVALIDHISTORICALNITROGENBUDGETS}
\usage{
fullVALIDHISTORICALNITROGENBUDGETS(aggregate = "glo")
}
\arguments{
\item{aggregate}{setting for aggregate, e.g.: TRUE, FALSE, "glo"}
}
\value{
magpie object
}
\description{
creates historical outputs of nitrogen budgets
}
\examples{
\dontrun{
fullVALIDHISTORICALNITROGENBUDGETS()
}
}
\seealso{
\code{\link{calcValidNitrogenPollution}}
}
\author{
Benjamin Leon Bodirsky
}
