% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullValidHistoricalNitrogenBudgets.R
\name{fullValidHistoricalNitrogenBudgets}
\alias{fullValidHistoricalNitrogenBudgets}
\title{fullValidHistoricalNitrogenBudgets}
\usage{
fullValidHistoricalNitrogenBudgets()
}
\value{
magpie object
}
\description{
creates historical outputs of nitrogen budgets
}
\examples{

\dontrun{ 
fullValidHistoricalNitrogenBudgets()
}
}
\seealso{
\code{\link{calcValidNitrogenPollution}}
}
\author{
Benjamin Leon Bodirsky
}
